refactor: decompose Assignment.js into facade with 7 sub-classes
- Create src/backend/AssignmentProcessor/Assignment/ folder with 00-06
  sub-class files + index.js facade
- Sub-classes: Serialisation, Factory, Rehydration, Timestamps,
  Submissions, AssessmentBase, LLMOrchestration
- Lazy getters on facade prototype for Object.create compatibility
- knownFields uses updatedAt (post-Section-2 name)
- Fix JSDoc: add constructor descriptions, @returns tags,
  defineLazySubclass documentation
- Fix security/detect-object-injection: use Object.entries +
  Object.defineProperty instead of bracket assignment
- All 1877 backend tests pass; lint clean (0 errors)
