--- src/backend/AssignmentProcessor/Assignment.js
@@ -1,658 +1,0 @@
-/**
- * Assignment Class
- *
- * Base class representing an assignment within a course, managing tasks and student submissions.
- */
-const INFO_TOAST_DURATION_SECONDS = 3;
-
-/**
- * Assignment class.
- */
-class Assignment {
-  /**
-   * Constructs an Assignment instance.
-   * @param {string} courseId - The ID of the course.
-   * @param {string} assignmentId - The ID of the assignment.
-   * @param {AssignmentDefinition|Object} assignmentDefinition - Embedded definition containing document type, reference/template IDs, and task metadata.
-   */
-  constructor(courseId, assignmentId, assignmentDefinition) {
-    this.courseId = courseId;
-    this.assignmentId = assignmentId;
-    this.assignmentName = this.fetchAssignmentName(courseId, assignmentId);
-    this.dueDate = null; //to be implemented later with the homework tracker.
-    // Timestamp for when this assignment was last updated. Use Date or null.
-    this.updatedAt = null;
-    // Timestamp from Google Classroom indicating when the assignment was created.
-    // Set by fetchAssignmentName() during construction.
-
-    // Embedded definition copy (source of truth for tasks, doc IDs, weighting, documentType)
-    this.assignmentDefinition = assignmentDefinition;
-
-    // New model: submissions array of StudentSubmission
-    this.submissions = []; // Array<StudentSubmission>
-    // Legacy studentTasks alias removed – callers must use this.submissions.
-    this.progressTracker = ProgressTracker.getInstance();
-    // Controllers may temporarily attach `assignment.students` while an assessment run is active
-    // to keep the hydrated roster handy. That property is transient and must never be persisted
-    // (see docs/developer/DATA_SHAPES.md and rehydration.md).
-    this._hydrationLevel = 'full';
-  }
-
-  /**
-   * Serialises this Assignment to a plain JSON-friendly object.
-   * Dates are converted to ISO strings. If TaskDefinition or StudentSubmission provide toJSON, those are used.
-   * progressTracker is intentionally not serialised (singleton/session-specific).
-   * @returns {object} Assignment data with course/assignment IDs, dates, definition, and submissions.
-   */
-  toJSON() {
-    const definitionJson = this.assignmentDefinition?.toJSON
-      ? this.assignmentDefinition.toJSON()
-      : this.assignmentDefinition;
-
-    const submissions = (this.submissions || []).map((sub) => {
-      if (sub && typeof sub.toJSON === 'function') return sub.toJSON();
-      // Fallback serialization for StudentSubmission-like objects
-      const out = {};
-      if (sub && typeof sub === 'object') {
-        out.studentId = sub.studentId || sub.userId || null;
-        if ('documentId' in sub) out.documentId = sub.documentId;
-        if ('score' in sub) out.score = sub.score;
-        if ('feedback' in sub) out.feedback = sub.feedback;
-        if (sub.updatedAt instanceof Date && !Number.isNaN(sub.updatedAt.getTime()))
-          out.updatedAt = sub.updatedAt.toISOString();
-        else if (sub.updatedAt) out.updatedAt = sub.updatedAt;
-        // copy any other enumerable properties (non-enumerable like methods are ignored)
-        Object.keys(sub).forEach((k) => {
-          if (!Object.hasOwn(out, k)) out[k] = sub[k];
-        });
-      }
-      return out;
-    });
-
-    return {
-      courseId: this.courseId,
-      assignmentId: this.assignmentId,
-      assignmentName: this.assignmentName,
-      dueDate: this.dueDate ? this.dueDate.toISOString() : null,
-      updatedAt: this.updatedAt ? this.updatedAt.toISOString() : null,
-      createdAt: this.createdAt.toISOString(),
-      ...this._extractFullDefinitionFields(definitionJson),
-      submissions,
-      assignmentDefinition: definitionJson || this.assignmentDefinition,
-    };
-  }
-
-  /**
-   * Extracts full definition fields from the serialised definition object.
-   * Used by toJSON to include complete definition data (document type, IDs, and tasks).
-   * @param {object} definitionJson - The serialised definition object.
-   * @returns {object} Full definition fields including documentType, IDs, and tasks.
-   */
-  _extractFullDefinitionFields(definitionJson) {
-    return {
-      documentType: definitionJson?.documentType ?? null,
-      referenceDocumentId: definitionJson?.referenceDocumentId ?? null,
-      templateDocumentId: definitionJson?.templateDocumentId ?? null,
-      tasks: definitionJson?.tasks ?? null,
-    };
-  }
-
-  /**
-   * Extracts minimal root fields for partial definitions.
-   * Only includes documentType (for routing); omits doc IDs and tasks.
-   * @param {object} definitionJson - The serialised definition object.
-   * @returns {object} Minimal root fields with only documentType.
-   */
-  _extractPartialRootFields(definitionJson) {
-    return {
-      documentType: definitionJson?.documentType ?? null,
-    };
-  }
-
-  /**
-   * Produces a lightweight JSON payload with heavy artifact fields redacted.
-   * @returns {object} Assignment data with redacted definition.
-   */
-  toPartialJSON() {
-    const definitionJson = this.assignmentDefinition.toPartialJSON();
-
-    const partialSubmissions = (this.submissions || []).map((submission) =>
-      submission.toPartialJSON()
-    );
-
-    return {
-      courseId: this.courseId,
-      assignmentId: this.assignmentId,
-      assignmentName: this.assignmentName,
-      dueDate: this.dueDate ? this.dueDate.toISOString() : null,
-      updatedAt: this.updatedAt ? this.updatedAt.toISOString() : null,
-      createdAt: this.createdAt.toISOString(),
-      ...this._extractPartialRootFields(definitionJson),
-      submissions: partialSubmissions,
-      assignmentDefinition: definitionJson,
-    };
-  }
-
-  /**
-   * Factory method to create the correct Assignment subclass based on documentType.
-   * @param {AssignmentDefinition|Object} assignmentDefinition - Embedded definition containing docType and task metadata.
-   * @param {string} courseId - The ID of the course.
-   * @param {string} assignmentId - The ID of the assignment.
-   * @returns {Assignment} Instance of appropriate subclass (SlidesAssignment or SheetsAssignment).
-   * @throws {Error} If documentType is invalid or unknown.
-   */
-  static create(assignmentDefinition, courseId, assignmentId) {
-    if (!assignmentDefinition?.documentType) {
-      throw new TypeError(
-        'assignmentDefinition with documentType is required to create Assignment.'
-      );
-    }
-
-    const type = assignmentDefinition.documentType.toUpperCase();
-
-    if (type === 'SLIDES') {
-      return new SlidesAssignment(courseId, assignmentId, assignmentDefinition);
-    }
-
-    if (type === 'SHEETS') {
-      return new SheetsAssignment(courseId, assignmentId, assignmentDefinition);
-    }
-
-    throw new Error(
-      `Unknown documentType: ${assignmentDefinition.documentType}. Valid types are 'SLIDES' or 'SHEETS'. See docs/developer/DATA_SHAPES.md for details.`
-    );
-  }
-
-  /**
-   * Internal helper to restore base Assignment fields from JSON data.
-   * Used by both base Assignment and subclass fromJSON methods.
-   * @param {object} data - JSON data object.
-   * @returns {Assignment} Assignment instance with base fields populated.
-   */
-  static _baseFromJSON(data) {
-    if (!data || typeof data !== 'object')
-      throw new Error('Invalid data supplied to Assignment._baseFromJSON');
-
-    if (!data.courseId || !data.assignmentId) {
-      throw new Error('courseId and assignmentId are required fields in Assignment data');
-    }
-
-    const inst = Object.create(Assignment.prototype);
-    inst.courseId = data.courseId;
-    inst.assignmentId = data.assignmentId;
-    inst.assignmentName = data.assignmentName || `Assignment ${data.assignmentId}`;
-    inst.dueDate = data.dueDate ? new Date(data.dueDate) : null;
-    inst.updatedAt = data.updatedAt ? new Date(data.updatedAt) : null;
-    if (!data.createdAt) {
-      throw new Error(`createdAt is required to deserialize Assignment ${data.assignmentId}`);
-    }
-    inst.createdAt = new Date(data.createdAt);
-    inst.assignmentDefinition = data.assignmentDefinition
-      ? AssignmentDefinition.fromJSON(data.assignmentDefinition)
-      : null;
-    inst.submissions = [];
-    // Do not set transient hydration marker here — remain absent/undefined so
-    // that deserialized objects don't claim a persisted hydration level.
-
-    if (Array.isArray(data.submissions)) {
-      data.submissions.forEach((subObject) => Assignment._rehydrateSubmission(inst, subObject));
-    }
-
-    // restore progress tracker singleton if available
-    inst.progressTracker = ProgressTracker.getInstance();
-
-    // Copy any additional fields that aren't already handled (e.g., referenceDocumentId, templateDocumentId for graceful degradation)
-    const knownFields = new Set([
-      'courseId',
-      'assignmentId',
-      'assignmentName',
-      'dueDate',
-      'updatedAt',
-      'createdAt',
-      'assignmentDefinition',
-      'submissions',
-      'students', // Transient, don't restore
-      'progressTracker', // Transient, don't restore
-      '_hydrationLevel', // Transient, don't restore
-    ]);
-    Object.keys(data).forEach((key) => {
-      if (!knownFields.has(key)) {
-        inst[key] = data[key];
-      }
-    });
-
-    // Do not populate `inst.students` here; any roster data should be sourced from ABClass at runtime
-    // and treated as ephemeral to avoid duplicate persistence.
-
-    return inst;
-  }
-
-  /**
-   * Rehydrates a submission object from JSON data and adds it to the assignment's submissions array.
-   * Handles StudentSubmission deserialization or falls back to plain object reconstruction.
-   * @param {Assignment} inst - The Assignment instance to add the submission to.
-   * @param {object} subObject - The submission object to rehydrate.
-   */
-  static _rehydrateSubmission(inst, subObject) {
-    const identifier = subObject && (subObject.studentId || subObject.userId);
-
-    try {
-      if (
-        typeof StudentSubmission !== 'undefined' &&
-        typeof StudentSubmission.fromJSON === 'function'
-      ) {
-        const submission = StudentSubmission.fromJSON(subObject);
-        inst.submissions.push(submission);
-        return;
-      }
-    } catch (error) {
-      ABLogger.getInstance().warn(
-        `StudentSubmission.fromJSON threw for studentId=${identifier}:`,
-        error
-      );
-    }
-
-    try {
-      const submission = new StudentSubmission(
-        identifier || null,
-        inst.assignmentId,
-        subObject.documentId || null,
-        subObject.studentName || subObject.name || null
-      );
-      Object.keys(subObject || {}).forEach((key) => {
-        if (key === 'updatedAt' && subObject.updatedAt) {
-          submission.updatedAt =
-            subObject.updatedAt instanceof Date
-              ? subObject.updatedAt
-              : new Date(subObject.updatedAt);
-          return;
-        }
-        submission[key] = subObject[key];
-      });
-      inst.submissions.push(submission);
-    } catch (error_) {
-      ABLogger.getInstance().warn(
-        `StudentSubmission reconstruction failed for studentId=${identifier}:`,
-        error_
-      );
-      const raw = { ...subObject };
-      if (raw.updatedAt) {
-        if (raw.updatedAt instanceof Date) raw.updatedAt = raw.updatedAt.toISOString();
-        else if (Validate.isString(raw.updatedAt)) {
-          const parsed = new Date(raw.updatedAt);
-          if (!Number.isNaN(parsed.getTime())) raw.updatedAt = parsed.toISOString();
-        }
-      }
-      inst.submissions.push(raw);
-    }
-  }
-
-  /**
-   * Polymorphic deserialisation routing based on documentType field.
-   * Routes to appropriate subclass fromJSON or creates base Assignment for legacy data.
-   * @param {object} data - JSON data object.
-   * @returns {Assignment} Instance of appropriate class (SlidesAssignment, SheetsAssignment, or base Assignment).
-   */
-  static fromJSON(data) {
-    if (!data || typeof data !== 'object')
-      throw new Error('Invalid data supplied to Assignment.fromJSON');
-
-    if (!data.courseId || !data.assignmentId) {
-      throw new Error('courseId and assignmentId are required fields in Assignment data');
-    }
-
-    if (!data.assignmentDefinition) {
-      if (!data.documentType) {
-        ProgressTracker.getInstance().logAndThrowError(
-          `Assignment data missing documentType for courseId=${data.courseId}, assignmentId=${data.assignmentId}`,
-          { data }
-        );
-      }
-
-      data.assignmentDefinition = new AssignmentDefinition({
-        primaryTitle: data.assignmentName || `Assignment ${data.assignmentId}`,
-        primaryTopic: data.assignmentName || 'Assignment',
-        documentType: data.documentType,
-        referenceDocumentId: data.referenceDocumentId,
-        templateDocumentId: data.templateDocumentId,
-        tasks: 'tasks' in data ? data.tasks : {},
-        referenceLastModified: data.referenceLastModified ?? null,
-        templateLastModified: data.templateLastModified ?? null,
-      }).toJSON();
-    }
-
-    const documentType = data.assignmentDefinition.documentType;
-
-    if (!documentType || typeof documentType !== 'string') {
-      ProgressTracker.getInstance().logAndThrowError(
-        `Assignment data missing documentType for courseId=${data.courseId}, assignmentId=${data.assignmentId}`,
-        { data }
-      );
-    }
-
-    const type = documentType.toUpperCase();
-
-    if (type === 'SLIDES') {
-      return SlidesAssignment.fromJSON(data);
-    }
-
-    if (type === 'SHEETS') {
-      return SheetsAssignment.fromJSON(data);
-    }
-
-    ProgressTracker.getInstance().logAndThrowError(
-      `Unknown assignment documentType '${documentType}' for courseId=${data.courseId}, assignmentId=${data.assignmentId}`,
-      { documentType: documentType, data }
-    );
-  }
-  /**
-   * Fetches the assignment name from Google Classroom.
-   * @param {string} courseId - The ID of the course.
-   * @param {string} assignmentId - The ID of the assignment.
-   * @returns {string} The name/title of the assignment.
-   */
-  fetchAssignmentName(courseId, assignmentId) {
-    const courseWork = Classroom.Courses.CourseWork.get(courseId, assignmentId);
-    if (!courseWork.creationTime) {
-      throw new Error(
-        `Google Classroom assignment ${assignmentId} (course ${courseId}) has no creationTime.`
-      );
-    }
-    this.createdAt = new Date(courseWork.creationTime);
-    return courseWork.title || `Assignment ${assignmentId}`;
-  }
-
-  /**
-   * Updates the updatedAt timestamp to the current date/time.
-   * Call this whenever the assignment is modified in-memory.
-   * @returns {Date} The new updatedAt value.
-   */
-  touchUpdated() {
-    // Use setUpdatedAt to centralize validation/copying behaviour.
-    return this.setUpdatedAt(new Date());
-  }
-
-  /**
-   * Returns the updatedAt Date or null if not set.
-   * @returns {Date|null} The updatedAt Date or null.
-   */
-  getUpdatedAt() {
-    return this.updatedAt || null;
-  }
-
-  /**
-   * Sets the updatedAt timestamp from a JavaScript Date object (or null to clear).
-   * The method copies the provided Date to avoid external mutation.
-   * @param {Date|null} date - The Date to set, or null to clear the timestamp.
-   * @returns {Date|null} The stored Date instance or null.
-   * @throws {TypeError} If the provided value is not a Date or null.
-   */
-  setUpdatedAt(date) {
-    if (date === null) {
-      this.updatedAt = null;
-      return null;
-    }
-    if (!(date instanceof Date) || Number.isNaN(date.getTime())) {
-      throw new TypeError('setUpdatedAt expects a valid Date or null');
-    }
-    // store a copy to avoid outside mutation
-    this.updatedAt = new Date(date);
-    return this.updatedAt;
-  }
-
-  /**
-   * Returns the createdAt Date.
-   * @returns {Date} The createdAt Date.
-   */
-  getCreatedAt() {
-    return this.createdAt;
-  }
-
-  /**
-   * Sets the createdAt timestamp from a JavaScript Date object.
-   * The method copies the provided Date to avoid external mutation.
-   * @param {Date} date - The Date to set.
-   * @returns {Date} The stored Date instance.
-   * @throws {TypeError} If the provided value is not a valid Date.
-   */
-  setCreatedAt(date) {
-    if (!(date instanceof Date) || Number.isNaN(date.getTime())) {
-      throw new TypeError('setCreatedAt expects a valid Date');
-    }
-    this.createdAt = new Date(date);
-    return this.createdAt;
-  }
-
-  /**
-   * Adds a student to the assignment.
-   * @param {Student} student - The Student instance to add.
-   * @returns {StudentSubmission|null} The created StudentSubmission, or null if studentId is not resolvable.
-   */
-  addStudent(student) {
-    // Expect student object with id
-    const studentId = student.id || student.studentId || student.userId;
-    if (!studentId) {
-      ABLogger.getInstance().warn('addStudent called without resolvable studentId');
-      return null;
-    }
-    // Avoid duplicates
-    const existing = this.submissions.find((s) => s.studentId === studentId);
-    if (existing) return existing;
-    const studentName =
-      student && (student.name || student.studentName || student.fullName)
-        ? student.name || student.studentName || student.fullName
-        : null;
-    const submission = new StudentSubmission(studentId, this.assignmentId, null, studentName);
-    // Attach original student metadata for any legacy code (non-persisted)
-    submission._legacyStudent = student;
-    this.submissions.push(submission);
-    return submission;
-  }
-
-  /**
-   * Processes a single attachment for a student's submission.
-   * Separates logic to reduce cognitive complexity in the parent method.
-   * @param {object} attachment - The attachment object from Classroom submission.
-   * @param {string} studentId - The Google Classroom student ID.
-   * @param {string} mimeType - The expected MIME type to validate against.
-   */
-  _processAttachmentForSubmission(attachment, studentId, mimeType) {
-    const driveFileId = attachment?.driveFile?.id;
-    if (!driveFileId) {
-      ABLogger.getInstance().info(
-        `Attachment for student ID ${studentId} is not a Drive File or lacks a valid ID.`
-      );
-      return;
-    }
-
-    try {
-      // Fetch the Drive file using DriveApp
-      const file = DriveApp.getFileById(driveFileId);
-      const fileMimeType = file.getMimeType();
-      if (this.isValidMimeType(fileMimeType, mimeType)) {
-        const documentId = driveFileId;
-        // New model: submissions array holds StudentSubmission objects with studentId
-        const submissionObject = this.submissions.find((sub) => sub.studentId === studentId);
-        if (submissionObject) {
-          submissionObject.documentId = documentId;
-          // Keep updatedAt coherent if method exists
-          if (typeof submissionObject.touchUpdated === 'function') submissionObject.touchUpdated();
-        } else {
-          ABLogger.getInstance().info(`No matching submission found for student ID: ${studentId}`);
-        }
-      } else {
-        ABLogger.getInstance().info(
-          `Attachment with Drive File ID ${driveFileId} is not a supported document (MIME type: ${fileMimeType}).`
-        );
-      }
-    } catch (fileError) {
-      ABLogger.getInstance().error(`Error fetching Drive file with ID ${driveFileId}:`, fileError);
-    }
-  }
-
-  /**
-   * Fetches and assigns submitted Google Drive documents for each student, filtered by the provided MIME type.
-   * @param {string} mimeType - The Google Drive MIME type to filter for (e.g., MimeType.GOOGLE_SLIDES, MimeType.GOOGLE_SHEETS).
-   * @returns {void}
-   */
-  fetchSubmittedDocumentsByMimeType(mimeType) {
-    try {
-      // Fetch all student submissions for the specific assignment
-      const response = Classroom.Courses.CourseWork.StudentSubmissions.list(
-        this.courseId,
-        this.assignmentId
-      );
-      const submissions = response.studentSubmissions;
-
-      if (!submissions || submissions.length === 0) {
-        ABLogger.getInstance().info(`No submissions found for assignment ID: ${this.assignmentId}`);
-        return;
-      }
-
-      submissions.forEach((submission) => {
-        const studentId = submission.userId; // Google Classroom Student ID (string)
-        const attachments = submission.assignmentSubmission?.attachments;
-
-        if (attachments && attachments.length > 0) {
-          attachments.forEach((attachment) =>
-            this._processAttachmentForSubmission(attachment, studentId, mimeType)
-          );
-        } else {
-          ABLogger.getInstance().info(`No attachments found for student ID: ${studentId}`);
-        }
-      });
-    } catch (error) {
-      ABLogger.getInstance().error(
-        `Error fetching submissions for assignment ID ${this.assignmentId}:`,
-        error
-      );
-    }
-  }
-
-  /**
-   * Validates if the file's MIME type matches the expected type.
-   * @param {string} fileMimeType - The MIME type of the file from Drive.
-   * @param {string} expectedMimeType - The expected Google MIME type.
-   * @returns {boolean} True if valid, false otherwise.
-   */
-  isValidMimeType(fileMimeType, expectedMimeType) {
-    return fileMimeType === expectedMimeType;
-  }
-
-  /**
-   * Fetches and assigns submitted documents for each student.
-   * This is a base method that should be implemented by subclasses.
-   * @returns {void}
-   */
-  fetchSubmittedDocuments() {
-    this._requireImplementation('fetchSubmittedDocuments');
-  }
-
-  /**
-   * Populates tasks from reference documents.
-   * This is a base method that should be implemented by subclasses.
-   * @returns {void}
-   */
-  populateTasks() {
-    this._requireImplementation('populateTasks');
-  }
-
-  /**
-   * Processes all student submissions by extracting responses.
-   * This is a base method that should be implemented by subclasses.
-   * @returns {void}
-   */
-  processAllSubmissions() {
-    this._requireImplementation('processAllSubmissions');
-  }
-
-  /**
-   * Generates an array of request objects ready to be sent to the LLM.
-   * @returns {Object[]} An array of request objects.
-   */
-  generateLLMRequests() {
-    return this._getLLMManager().generateRequestObjects(this);
-  }
-
-  /**
-   * Assesses student responses by interacting with the LLM.
-   * Generates LLM requests for all submissions and processes responses.
-   * @returns {void}
-   */
-  assessResponses() {
-    // Base Assignment only handles non-spreadsheet (text/table/image) via LLM
-    const manager = this._getLLMManager();
-    const requests = manager.generateRequestObjects(this);
-    if (!requests || requests.length === 0) {
-      Utils.toastMessage('No LLM requests to send.', 'Info', INFO_TOAST_DURATION_SECONDS);
-      return;
-    }
-    manager.processStudentResponses(requests, this);
-  }
-
-  /**
-   * Creates and returns an LLMRequestManager instance.
-   * Centralises construction to reduce duplicated logic across multiple methods.
-   * @returns {LLMRequestManager} A new LLMRequestManager instance.
-   */
-  _getLLMManager() {
-    return new LLMRequestManager();
-  }
-
-  /**
-   * Small helper used by base-class methods that must be implemented by subclasses.
-   * Centralising the throw here reduces the duplicated error message logic across
-   * multiple tiny abstract-style methods.
-   * @param {string} methodName - Name of the method that should be implemented
-   */
-  _requireImplementation(methodName) {
-    throw new Error(`${methodName} must be implemented by subclasses`);
-  }
-
-  /**
-   * Gets the tasks object from the assignment definition.
-   * @returns {Object|null} Task definitions keyed by task ID, or null if not set.
-   */
-  getTasks() {
-    return this.assignmentDefinition?.tasks ?? null;
-  }
-
-  /**
-   * Sets the tasks object in the assignment definition.
-   * @param {Object} tasks - Task definitions keyed by task ID.
-   * @returns {Object} The assigned tasks object.
-   */
-  setTasks(tasks) {
-    this.assignmentDefinition.tasks = tasks;
-    return tasks;
-  }
-
-  /**
-   * Gets the document type from the assignment definition.
-   * @returns {string|null} The document type (e.g., 'SLIDES', 'SHEETS'), or null if not set.
-   */
-  getDocumentType() {
-    return this.assignmentDefinition?.documentType ?? null;
-  }
-
-  /**
-   * Gets the reference document ID from the assignment definition.
-   * @returns {string|null} The reference document ID, or null if not set.
-   */
-  getReferenceDocumentId() {
-    return this.assignmentDefinition?.referenceDocumentId ?? null;
-  }
-
-  /**
-   * Gets the template document ID from the assignment definition.
-   * @returns {string|null} The template document ID, or null if not set.
-   */
-  getTemplateDocumentId() {
-    return this.assignmentDefinition?.templateDocumentId ?? null;
-  }
-}
-
-// Export for Node/Vitest environment (ignored in GAS runtime)
-if (typeof module !== 'undefined') {
-  module.exports = Assignment;
-}
