+++ tests/assignment/assignmentSerialisation.test.js
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import Assignment from '../../src/backend/AssignmentProcessor/Assignment.js';
+import Assignment from '../../src/backend/AssignmentProcessor/Assignment/index.js';
 import { AssignmentDefinition } from '../../src/backend/Models/AssignmentDefinition.js';
 import {
   createSlidesAssignment,
