+++ tests/helpers/modelFactories.js
@@ -7,7 +7,7 @@ const { TaskDefinition } = require('../../src/backend/Models/TaskDefinition.js')
 const { StudentSubmission } = require('../../src/backend/Models/StudentSubmission.js');
 const { ArtifactFactory } = require('../../src/backend/Models/Artifacts/index.js');
 const { AssignmentDefinition } = require('../../src/backend/Models/AssignmentDefinition.js');
-const Assignment = require('../../src/backend/AssignmentProcessor/Assignment.js');
+const Assignment = require('../../src/backend/AssignmentProcessor/Assignment/index.js');
 
 /**
  * Create a TaskDefinition for testing with sensible defaults
